Выносит cards в index.js
